+++ src/sdc/writer/_dpt.py
@@ -38,6 +38,7 @@ class DPTWriter(SplittableStreamWriter, InputBasedPlaceholderSupporter):
         self.output_dir = output_dir
         self.descending = descending
         self.locale = locale
+        self._writer = None
 
     def name(self) -> str:
         """
@@ -100,6 +101,8 @@ class DPTWriter(SplittableStreamWriter, InputBasedPlaceholderSupporter):
             self.descending = False
         if self.locale is None:
             self.locale = "en_US"
+        self._writer = SWriter()
+        self._writer.options = self._compile_options()
 
     def _compile_options(self) -> List[str]:
         """
@@ -120,9 +123,6 @@ class DPTWriter(SplittableStreamWriter, InputBasedPlaceholderSupporter):
 
         :param data: the data to write (single record or iterable of records)
         """
-        writer = SWriter()
-        writer.options = self._compile_options()
-
         for item in make_list(data):
             sub_dir = self.session.expand_placeholders(self.output_dir)
             if self.splitter is not None:
@@ -135,4 +135,4 @@ class DPTWriter(SplittableStreamWriter, InputBasedPlaceholderSupporter):
             path = os.path.join(sub_dir, item.spectrum_name)
             path = os.path.splitext(path)[0] + ".dpt"
             self.logger().info("Writing spectrum to: %s" % path)
-            writer.write([item.spectrum], path)
+            self._writer.write([item.spectrum], path)