+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changelog
 0.0.2 (????-??-??)
 ------------------
 
-- wai.spectralio readers now instantiate the reader in the `initialize` method
+- wai.spectralio-based readers now instantiate the wai.spectralio reader in the `initialize` method
+- wai.spectralio-based writers now instantiate the wai.spectralio writer in the `initialize` method
 
 
 0.0.1 (2025-06-27)